% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_alpha.R
\name{cal_alpha}
\alias{cal_alpha}
\title{Calculate alpha diversity metrics}
\usage{
cal_alpha(data, metrics = c("Observed", "Shannon", "Chao1", "Simpson"))
}
\arguments{
\item{data}{A phyloseq object}

\item{metrics}{A character vector of alpha diversity metrics to calculate (default: c("Observed", "Shannon", "Chao1", "Simpson"))}
}
\value{
A data.frame of alpha diversity values
}
\description{
Calculate alpha diversity metrics
}
