% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_global.R
\name{deseq_global}
\alias{deseq_global}
\title{Run DESeq2 time-series interaction test}
\usage{
deseq_global(
  physeq,
  time_var,
  group_var,
  taxrank = "Genus",
  padj_cutoff = 0.05
)
}
\arguments{
\item{physeq}{A phyloseq object (already aggregated if needed)}

\item{time_var}{Name of the time variable (user-defined)}

\item{group_var}{Name of the treatment group variable (user-defined)}

\item{taxrank}{Taxonomic rank for annotation (e.g. "Genus")}

\item{padj_cutoff}{Adjusted p-value threshold (default = 0.05)}
}
\value{
A list with the result table and ggplot object
}
\description{
Run DESeq2 time-series interaction test
}
