% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_metadata.R
\name{import_metadata}
\alias{import_metadata}
\title{Import metadata from CSV, TSV, TXT, or Excel}
\usage{
import_metadata(metadata_path)
}
\arguments{
\item{metadata_path}{Path to the metadata file}
}
\value{
A data.frame containing metadata
}
\description{
Import metadata from CSV, TSV, TXT, or Excel
}
