% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_phyloseq.R
\name{import_phyloseq}
\alias{import_phyloseq}
\title{Import phyloseq components and assign as global variables}
\usage{
import_phyloseq(phyloseq_path)
}
\arguments{
\item{phyloseq_path}{Path to the saved RDS phyloseq object}
}
\value{
NULL (assigns ps, taxonomy, table, metadata to global environment)
}
\description{
Import phyloseq components and assign as global variables
}
