% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse_tools.R
\name{phyloseq_to_se}
\alias{phyloseq_to_se}
\title{Convert phyloseq to SummarizedExperiment (auto transpose if needed)}
\usage{
phyloseq_to_se(ps)
}
\arguments{
\item{ps}{A phyloseq object}
}
\value{
A SummarizedExperiment object
}
\description{
This function converts a phyloseq object to a SummarizedExperiment object,
and automatically transposes the OTU table if taxa_are_rows is FALSE.
}
\examples{
se <- phyloseq_to_se(ps)
}
