% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phylum.R
\name{plot_phylum}
\alias{plot_phylum}
\title{Plot Mean Relative Abundance at Phylum Level}
\usage{
plot_phylum(data, group_vars, facet = NULL, x_var = NULL)
}
\arguments{
\item{data}{A \code{phyloseq} object containing taxonomic and sample metadata.}

\item{group_vars}{A character vector of metadata variables to group by
(e.g., \code{c("treatment", "timeline")}).}

\item{facet}{Optional. A metadata variable to facet by (e.g., \code{"timeline"}).}

\item{x_var}{Optional. A variable to use as the x-axis. Must be one of \code{group_vars}.
If not provided, the first variable in \code{group_vars} will be used.}
}
\value{
A \code{ggplot2} object showing stacked bar plots of mean phylum relative abundance.
}
\description{
This function plots the mean relative abundance of phyla in a \code{phyloseq} object
using stacked bar charts. It allows grouping by one or more metadata variables
and optional faceting.
}
\details{
This function:
\itemize{
\item Aggregates data at the Phylum level using \code{tax_glom}.
\item Converts counts to relative abundance.
\item Groups and summarizes mean relative abundance across the provided metadata variables.
\item Optionally facets the plot by an additional variable.
}
}
\examples{
\dontrun{
plot_phylum(data = ps, group_vars = c("treatment", "timeline"), x_var = "treatment", facet = "timeline")
}

}
\seealso{
\code{\link{plot_genus}}, \code{\link[phyloseq]{tax_glom}}, \code{\link[ggplot2]{ggplot}}
}
