% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse_tools.R
\name{run_lefse_pairwise}
\alias{run_lefse_pairwise}
\title{Run pairwise LEfSe analysis and return both result and plot}
\usage{
run_lefse_pairwise(se, classCol, groups, lda_threshold = 2)
}
\arguments{
\item{se}{A SummarizedExperiment object}

\item{classCol}{The name of the metadata column containing the group information}

\item{groups}{A character vector of 2 group levels to compare}

\item{lda_threshold}{LDA score cutoff}
}
\value{
A list with:
\item{res}{data.frame of LEfSe results}
\item{plot}{a ggplot2 object for LDA visualization}
}
\description{
This function runs LEfSe pairwise between two groups from a SummarizedExperiment,
and returns both the result table and the LDA plot.
}
\examples{
# out <- run_lefse_pairwise(se, classCol = "group", groups = c("A", "B"))
# print(out$plot); head(out$res)
}
